+++ __tests__/threeScene.spec.js
@@ -27,7 +27,6 @@ describe('Three scene', () => {
 
     checkAxisIndicatorDefaultState(scene);
 
-    expect(scene.animationFrame).toBeNull();
     expect(scene.layerSet).toBeInstanceOf(LayerSet);
     expect(scene.originalScene).toBeInstanceOf(Scene);
   });
@@ -68,5 +67,27 @@ describe('Three scene', () => {
     expect(scene.axisIndicator.camera).toBeInstanceOf(PerspectiveCamera);
     expect(scene.axisIndicator.canvas).toBeInstanceOf(HTMLCanvasElement);
     expect(scene.axisIndicator.renderer).not.toBe(null);
+  });
+
+  it('should stop animating and remove event handlers when destroyed', async () => {
+    const scene = new ThreeScene();
+    const el = document.createElement('div');
+    const resizeHandlerSpy = vi.spyOn(scene, 'onResize');
+
+    scene.initialize(el);
+
+    window.dispatchEvent(new Event('resize'));
+    window.dispatchEvent(new Event('resize'));
+
+    expect(scene.isAnimating).toBe(true);
+    expect(resizeHandlerSpy).toHaveBeenCalledTimes(3);
+
+    scene.destroy();
+
+    window.dispatchEvent(new Event('resize'));
+
+    expect(scene.isAnimating).toBe(false);
+    expect(resizeHandlerSpy).toHaveBeenCalledTimes(3);
+
   });
 });